# Configurable reward frequency. 2x times faster tests. No drawbacks.