disabled: make travis_rpc_checker  0.47s user 0.30s system 0% cpu **6:25.26** total
enabled:  make travis_rpc_checker  0.45s user 0.28s system 0% cpu **3:46.12** total